+++ build_web_compilers/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 4.0.12-wip
 
-- Bump the min sdk to 3.5.0.
 - Account for dartdevc snapshot in the Dart SDK changing to an AOT snapshot.
+- Support compiling to WebAssembly by using `dart2wasm` as a value for the
+  compiler option.
+- Bump the min sdk to 3.6.0-165.0.dev.
 
 ## 4.0.11